shaped up some comments and moved services
